feature-home: show upcoming renewals in the important-messages carousel
Per design feedback, informational reminders (upcoming renewals) should appear
in the same horizontal carousel as very important messages rather than being
dropped from home. Introduce a HomeNoticeCard model and a HomeNoticeCarousel
that renders both very important messages and renewal cards, add a
homeInformationalReminders split, and gate the section on either being present.
The To do list still shows only action-required reminders.

Also moves the carousel and card composables out of HomeDestination into a new
HomeNoticeCards file to keep the large file smaller.
